Adding the relationship between Products and Orders with use of endpoints (error handling done)

+++ src/main/java/com/alexdev/ecommercebackend/model/dto/OrderDTO.java
@@ -20,7 +20,7 @@ public class OrderDTO implements Serializable {
     private int id;
 
     @PositiveOrZero(message = "cannot be under 0")
-    private int ammount;
+    private double ammount;
 
     private String shippingAddress;
